app/src/main/java/com/udacity/gradle/builditbigger/GetJokeAsyncTask.java
package com.udacity.gradle.builditbigger;

import android.content.Context;
import android.os.AsyncTask;
import android.util.Pair;

import com.google.api.client.extensions.android.http.AndroidHttp;
import com.google.api.client.extensions.android.json.AndroidJsonFactory;
import com.google.api.client.googleapis.services.AbstractGoogleClientRequest;
import com.google.api.client.googleapis.services.GoogleClientRequestInitializer;
import com.udacity.gradle.builditbigger.backend.myApi.MyApi;

import java.io.IOException;

public class GetJokeAsyncTask extends AsyncTask< GetJokeAsyncTask.JokeListener, Void, String> {

    private static MyApi myApiService = null;
    private JokeListener jokeListener;

    @Override
    protected String doInBackground( JokeListener... params) {

        if(myApiService == null) {
           buildApiService();
        }

        this.jokeListener = params[0];

        try {
            return myApiService.getJoke().execute().getData();
        } catch (IOException e) {
            return e.getMessage();
        }
    }

    private void buildApiService() {
        MyApi.Builder builder = new MyApi.Builder(AndroidHttp.newCompatibleTransport(),
                new AndroidJsonFactory(), null)
                .setRootUrl("http://10.0.2.2:8080/_ah/api/")
                .setGoogleClientRequestInitializer(new GoogleClientRequestInitializer() {
                    @Override
                    public void initialize(AbstractGoogleClientRequest<?> abstractGoogleClientRequest) throws IOException {
                        abstractGoogleClientRequest.setDisableGZipContent(true);
                    }
                });

        myApiService = builder.build();
    }

    @Override
    protected void onPostExecute(String result) {
        jokeListener.receivedJoke(result);
    }

    public interface JokeListener {
        void receivedJoke(String joke);
    }
}